Factor out chunks into separate module
